elastic/requirements.txt
pytest==6.1.2
pytest-asyncio==0.12.0
aiohttp==3.7.2
aioredis==1.3.1
elasticsearch==7.15.2
python-dotenv==0.19.1
pydantic==1.8.2
requests==2.26.0